+++ src/main/java/com/unzer/payment/Resource.java
@@ -8,6 +8,7 @@ public interface Resource {
      * Returns resource unique id.
      * <br/>
      * For example: s-pay-21
+     *
      * @return resource id
      */
     String getId();
@@ -16,6 +17,7 @@ public interface Resource {
      * Returns the versioned URL of the resource.
      * <br/>
      * For example: /v1/payments/s-pay-21/charges/s-chg-3
+     *
      * @return resource URL
      */
     String getUrl();