+++ src/main/java/com/unzer/payment/models/googlepay/SignedKey.java
@@ -1,0 +1,9 @@
+package com.unzer.payment.models.googlepay;
+
+import lombok.Data;
+
+@Data
+public class SignedKey {
+    private String keyExpiration;
+    private String keyValue;
+}