+++ src/main/java/com/unzer/payment/models/googlepay/SignedMessage.java
@@ -1,0 +1,10 @@
+package com.unzer.payment.models.googlepay;
+
+import lombok.Data;
+
+@Data
+public class SignedMessage {
+    private String tag;
+    private String ephemeralPublicKey;
+    private String encryptedMessage;
+}