+++ src/test/java/com/unzer/payment/business/paymenttypes/PayUTest.java
@@ -27,12 +27,18 @@ class PayUTest {
     @Test
     void charge() {
         Unzer unzer = new Unzer(new HttpClientMock(), "s-private-key");
-        stubFor(post("/v1/payments/charges/").willReturn(
-                jsonResponse(getResponse("charge.json"), 200)));
-        stubFor(get("/v1/payments/s-pay-286").willReturn(
-                jsonResponse(getResponse("fetch-payment.json"), 200)));
-        stubFor(get("/v1/payments/s-pay-286/charges/s-chg-1").willReturn(
-                jsonResponse(getResponse("fetch-charge.json"), 200)));
+        stubFor(
+                post("/v1/payments/charges/")
+                        .willReturn(jsonResponse(getResponse("charge.json"), 200))
+        );
+        stubFor(
+                get("/v1/payments/s-pay-286")
+                        .willReturn(jsonResponse(getResponse("fetch-payment.json"), 200))
+        );
+        stubFor(
+                get("/v1/payments/s-pay-286/charges/s-chg-1")
+                        .willReturn(jsonResponse(getResponse("fetch-charge.json"), 200))
+        );
 
         Charge charge = unzer.charge((Charge) new Charge()
                 .setTypeId("s-pyu-6tg6nwdkrcdk")